Changes made on the plane from VA
git-svn-id: https://svn.nmap.org/trunk/nmap@481 e0a8ed71-7df4-0310-8962-fdc924857419

+++ global_structures.h
@@ -201,7 +201,7 @@ struct ops /* someone took struct options, <grrr> */ {
   int max_parallelism;
   int max_rtt_timeout;
   int min_rtt_timeout;
-  int host_timeout;
+  unsigned long host_timeout;
   int scan_delay;
   int initial_rtt_timeout;
